requirements.txt
feedparser>=6.0.10
anthropic>=0.25.0
python-dateutil>=2.8.2
markdown>=3.5.0
pyyaml>=6.0
stripe>=8.0.0
